clean(application): moved new interfaces to perception package
